+++ examples/1_function_test/test_list.h
@@ -130,7 +130,7 @@ private:
 		std::list<std::string> list_ref;
 
 		std::vector<int> vRoleIds;
-		const size_t COUNT = 1000;
+		const int COUNT = 1000;
 		for (auto i = 0; i < COUNT; i++) {
 			vRoleIds.push_back(i);
 		}
